fix(task-scheduler): change sql request for optimisation

+++ dossierfacile-task-scheduler/src/main/java/fr/dossierfacile/scheduler/tasks/garbagecollection/GarbageSequenceEntity.java
@@ -26,7 +26,7 @@ public class GarbageSequenceEntity {
     GarbageSequenceName name;
 
     @Column
-    int value;
+    long value;
 
     @Column
     LocalDateTime lastUpdateDate;
